Move common pkg to v2 since its v2 spec schema
