+++ pkg/devfile/parser/data/v2/common/component_helper.go
