+++ pkg/devfile/parser/data/v2/common/component_helper_test.go
