+++ pkg/devfile/parser/data/v2/common/project_helper.go
