+++ pkg/devfile/parser/data/v2/volumes.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	v1 "github.com/devfile/api/pkg/apis/workspaces/v1alpha2"
-	"github.com/devfile/parser/pkg/devfile/parser/data/common"
+	"github.com/devfile/parser/pkg/devfile/parser/data/v2/common"
 )
 
 // AddVolume adds the volume to the devFile and mounts it to all the container components
